fix to work around provider* functions getting called as part of loading the Routes config
