+++ libs/smart-ngrx/src/mark-and-delete/mark-and-delete-effect/mark-and-delete-entity.function.ts
@@ -20,7 +20,7 @@ export function markAndDeleteEntity([feature, entity]: [
   const garbageCollectRowIds: string[] = [];
   const markDirtyRowIds: string[] = [];
   const now = Date.now();
-  const runInterval = getGlobalMarkAndDeleteInit().runInterval;
+  const runInterval = getGlobalMarkAndDeleteInit().runInterval!;
   for (const [key, value] of entityMap) {
     if (
       0 === featureInit.removeTime &&
