+++ libs/smart-ngrx/src/mark-and-delete/mark-and-delete-effect/process-mark-and-delete.function.ts
@@ -37,6 +37,6 @@ export function processMarkAndDelete(
         store!.dispatch(entityAction.markDirty({ ids: markDirtyRowIds }));
       }
     },
-    { timeout: getGlobalMarkAndDeleteInit().runInterval - 100 },
+    { timeout: getGlobalMarkAndDeleteInit().runInterval! - 100 },
   );
 }
